+++ bioconda_utils/cli.py
@@ -9,7 +9,7 @@ Bioconda Utils Command Line Interface
 import warnings
 from bioconda_utils import bulk
 
-from bioconda_utils.artifacts import upload_pr_artifacts
+from bioconda_utils.artifacts import UploadResult, upload_pr_artifacts
 from bioconda_utils.skiplist import Skiplist
 from bioconda_utils.build_failure import BuildFailureRecord, collect_build_failure_dataframe
 warnings.filterwarnings("ignore", message="numpy.dtype size changed")
@@ -56,7 +56,7 @@ def enable_logging(default_loglevel='info', default_file_loglevel='debug'):
         @arg('--logfile', help="Write log to file")
         @arg('--logfile-level', help="Log level for log file")
         @arg('--log-command-max-lines', help="Limit lines emitted for commands executed")
-        @utils.wraps(func)
+        @utils.wraps(func, hide_wrapped=True)
         def wrapper(*args, loglevel=default_loglevel, logfile=None,
                     logfile_level=default_file_loglevel,
                     log_command_max_lines=None, **kwargs):
@@ -125,7 +125,7 @@ def recipe_folder_and_config(allow_missing_for=None):
                                  for field in allow_missing_for or []]
         except ValueError:
             sys.exit(f"Function {func} must have 'recipe_folder' and 'config' args")
-        @arg('recipe_folder', nargs='?',
+        @arg('recipe-folder', nargs='?',
              help='Path to folder containing recipes (default: recipes/)')
         @arg('config', nargs='?',
              help='Path to Bioconda config (default: config.yml)')
@@ -238,23 +238,25 @@ def duplicates(config,
         check_fields += ['build']
 
     def remove_package(spec):
-        fn = '{}-{}-{}.tar.bz2'.format(*spec)
-        name, version = spec[:2]
-        subcmd = [
-            'remove', '-f',
-            '{channel}/{name}/{version}/{fn}'.format(
-                name=name, version=version, fn=fn, channel=our_channel
-            )
-        ]
-        if dryrun:
-            logger.info(" ".join([utils.bin_for('anaconda')] + subcmd))
-        else:
-            token = os.environ.get('ANACONDA_TOKEN')
-            if token is None:
-                token = []
+        for ext in (".tar.bz2", ".conda"):
+            name, version = spec[:2]
+            dist = '{}-{}-{}'.format(*spec)
+            fn = f"{dist}{ext}"
+            subcmd = [
+                'remove', '-f',
+                '{channel}/{name}/{version}/{fn}'.format(
+                    name=name, version=version, fn=fn, channel=our_channel
+                )
+            ]
+            if dryrun:
+                logger.info(" ".join([utils.bin_for('anaconda')] + subcmd))
             else:
-                token = ['-t', token]
-            logger.info(utils.run([utils.bin_for('anaconda')] + token + subcmd, mask=[token]).stdout)
+                token = os.environ.get('ANACONDA_TOKEN')
+                if token is None:
+                    token = []
+                else:
+                    token = ['-t', token]
+                logger.info(utils.run([utils.bin_for('anaconda')] + token + subcmd, mask=[token]).stdout)
 
     # packages in our channel
     repodata = utils.RepoData()
@@ -436,6 +438,7 @@ from environment, even after successful build and test.''')
 @arg("--skiplist-leafs", action="store_true", help="Skiplist leaf recipes (i.e. ones that are not depended on by any other recipes) that fail to build.")
 @arg('--disable-live-logs', action='store_true', help="Disable live logging during the build process")
 @arg('--exclude', nargs='+', help='Packages to exclude during this run')
+@arg('--subdag-depth', type=int, help="Number of levels of root nodes to skip. (Optional, and only if using n_workers)")
 @enable_logging()
 def build(recipe_folder, config, packages="*", git_range=None, testonly=False,
           force=False, docker=None, mulled_test=False, build_script_template=None,
@@ -447,7 +450,8 @@ def build(recipe_folder, config, packages="*", git_range=None, testonly=False,
           record_build_failures=False,
           skiplist_leafs=False,
           disable_live_logs=False,
-          exclude=None):
+          exclude=None,
+          subdag_depth=None):
     cfg = utils.load_config(config)
     setup = cfg.get('setup', None)
     if setup:
@@ -510,6 +514,7 @@ def build(recipe_folder, config, packages="*", git_range=None, testonly=False,
                             skiplist_leafs=skiplist_leafs,
                             live_logs=(not disable_live_logs),
                             exclude=exclude,
+                            subdag_depth=subdag_depth
                             )
     exit(0 if success else 1)
 
@@ -524,7 +529,7 @@ def build(recipe_folder, config, packages="*", git_range=None, testonly=False,
 @arg('--dryrun', action='store_true', help='''Do not actually upload anything.''')
 @arg('--fallback', choices=['build', 'ignore'], default='build', help="What to do if no artifacts are found in the PR.")
 @arg('--quay-upload-target', help="Provide a quay.io target to push docker images to.")
-@arg('--artifact-source', choices=['azure', 'circleci'], default='azure', help="Application hosting build artifacts (e.g., Azure or Circle CI).")
+@arg('--artifact-source', choices=['azure', 'circleci','github-actions'], default='azure', help="Application hosting build artifacts (e.g., Azure, Circle CI, or GitHub Actions).")
 @enable_logging()
 def handle_merged_pr(
     recipe_folder,
@@ -538,10 +543,12 @@ def handle_merged_pr(
 ):
     label = os.getenv('BIOCONDA_LABEL', None) or None
 
-    success = upload_pr_artifacts(
-        config, repo, git_range[1], dryrun=dryrun, mulled_upload_target=quay_upload_target, label=label, artifact_source=artifact_source
+    res = upload_pr_artifacts(
+        config, repo, git_range[1], dryrun=dryrun,
+        mulled_upload_target=quay_upload_target, label=label,
+        artifact_source=artifact_source
     )
-    if not success and fallback == 'build':
+    if res == UploadResult.NO_ARTIFACTS and fallback == 'build':
         success = build(
             recipe_folder,
             config,
@@ -551,6 +558,8 @@ def handle_merged_pr(
             mulled_test=True,
             label=label,
         )
+    else:
+        success = res != UploadResult.FAILURE
     exit(0 if success else 1)
 
 @recipe_folder_and_config()
@@ -756,7 +765,7 @@ def dependent(recipe_folder, config, restrict=False,
      is specified, then all packages in a given bioconductor release will be
      created/updated (--force is then implied).''')
 @recipe_folder_and_config()
-@arg('bioc_data_packages', nargs='?',
+@arg('bioc-data-packages', nargs='?',
      help='''Path to folder containing the recipe for the bioconductor-data-packages
      (default: recipes/bioconductor-data-packages)''')
 @arg('--versioned', action='store_true', help='''If specified, recipe will be
@@ -847,8 +856,6 @@ def clean_cran_skeleton(recipe, no_windows=False):
     cran_skeleton.clean_skeleton_files(recipe, no_windows=no_windows)
 
 
-@arg('recipe_folder', help='Path to recipes directory')
-@arg('config', help='Path to yaml file specifying the configuration')
 @recipe_folder_and_config()
 @arg('--packages', nargs="+",
      help='Glob(s) for package[s] to scan. Can be specified more than once')