+++ bioconda_utils/docker_utils.py
@@ -83,7 +83,7 @@ logger = logging.getLogger(__name__)
 # filled in here.
 #
 BUILD_SCRIPT_TEMPLATE = \
-"""
+r"""
 #!/bin/bash
 set -eo pipefail
 
@@ -94,23 +94,32 @@ set -eo pipefail
 # will exist in the container but will be empty.  Channels expect at least
 # a linux-64/linux-aarch64 and noarch directory within that directory, so we
 # make sure it exists before adding the channel.
-mkdir -p {self.container_staging}/linux-64
-mkdir -p {self.container_staging}/linux-aarch64
-mkdir -p {self.container_staging}/noarch
-touch {self.container_staging}/noarch/repodata.json
+# Also ensure conda-build's local channel directory exists the same way.
+for local_channel in '/opt/conda/conda-bld' '{self.container_staging}'; do
+  mkdir -p "${{local_channel}}"/linux-64
+  mkdir -p "${{local_channel}}"/linux-aarch64
+  mkdir -p "${{local_channel}}"/noarch
+  conda index "${{local_channel}}"
+done
 conda config --add channels file://{self.container_staging} 2> >(
     grep -vF "Warning: 'file://{self.container_staging}' already in 'channels' list, moving to the top" >&2
 )
 
+# Pass on conda_pkg_format ("2" for .conda instead of .tar.bz2) from host's conda-build config.
+test -n '{self.conda_pkg_format}' && conda config --set conda_build.pkg_format '{self.conda_pkg_format}'
+
 # The actual building...
 # we explicitly point to the meta.yaml, in order to keep
 # conda-build from building all subdirectories
-conda mambabuild -c file://{self.container_staging} {self.conda_build_args} {self.container_recipe}/meta.yaml 2>&1
+conda-build -c file://{self.container_staging} {self.conda_build_args} {self.container_recipe}/meta.yaml 2>&1
 
 # copy all built packages to the staging area
-cp /opt/conda/conda-bld/*/*.tar.bz2 {self.container_staging}/{arch}
+find /opt/conda/conda-bld \
+  -name src_cache -prune -o \
+  -type f \( -name '*.tar.bz2' -o -name '*.conda' \) -print0 |
+  xargs -0 -- cp -t '{self.container_staging}/{arch}' --
 #While technically better, this is slower and more prone to breaking
-#cp `conda mambabuild {self.conda_build_args} {self.container_recipe}/meta.yaml --output | grep tar.bz2` {self.container_staging}/{arch}
+#cp `conda-build {self.conda_build_args} {self.container_recipe}/meta.yaml --output | grep -e '\.tar\.bz2$' -e '\.conda$')` {self.container_staging}/{arch}
 conda index {self.container_staging}
 # Ensure permissions are correct on the host.
 HOST_USER={self.user_info[uid]}
@@ -147,18 +156,6 @@ class DockerCalledProcessError(sp.CalledProcessError):
 class DockerBuildError(Exception):
     pass
 
-
-
-def get_host_conda_bld():
-    """
-    Identifies the conda-bld directory on the host.
-
-    Assumes that conda-build is installed.
-    """
-    # v0.16.2: this used to have a side effect, calling conda build purge
-    # hopefully, it's not actually needed.
-    build_conf = utils.load_conda_build_config()
-    return build_conf.build_folder
 
 
 class RecipeBuilder(object):
@@ -280,7 +277,11 @@ class RecipeBuilder(object):
         self.container_recipe = container_recipe
         self.container_staging = container_staging
 
-        self.host_conda_bld = get_host_conda_bld()
+        conda_build_config = utils.load_conda_build_config()
+        # Identify conda-bld directory on the host.
+        self.host_conda_bld = conda_build_config.croot
+        # Pass on config to choose wheter to build .tar.bz2 or .conda format.
+        self.conda_pkg_format = conda_build_config.conda_pkg_format or ""
 
         if use_host_conda_bld:
             self.pkg_dir = self.host_conda_bld